Reworked how we read the manifest files and thus what we sync
We no longer flatten the manifests into one level for each table and
instead handle individual dumps separately. This is necessary as
previously the tap assumed the `'incremental'` value was the same for all
dumps of a stream, but it can change.

This changes how we select what files to sync. Previously if we saw a
single dump with an `'incremental' = false` we would only sync the
files from the latest dump. Now we look backwards through the manifests to
find the latest `'incremental = false'`, and choose the max between that
and our bookmarked dump_id to determine which dumps to sync.