Enable AWS access key as alternative auth method
Previously this only allowed for IAM role access to the s3 bucket.
This commit allows for an alternative, with s3 access coming from an
account via providing aws_access_key_id and aws_secret_access_key to
the config. If access key pair is provided, it overrides an IAM role
information provided.